running with single/multi with/out save, need to test n_copies

+++ lab_space/experiment.py
@@ -50,7 +50,8 @@ class Experiment():
         
         self._log.warn("RunExperiment Init, perform " + str(expt_config["n_trials"]) + " trials across " + str(expt_config["n_processes"]) + " processes")
 
-        self.__lock = Lock()
+        global expt_lock
+        expt_lock = Lock()
         
         self._trial_config = []
         self._expt_config = {}
@@ -80,7 +81,7 @@ class Experiment():
             self._expt_config["save_file"] = None
         if "clear_save" not in self._expt_config:
             self._expt_config["clear_save"] = False
-        if self._expt_config["clear_save"] and os.path.isfile(self._expt_config["save_file"]):
+        if self._expt_config["clear_save"] and self._expt_config["save_file"] is not None and os.path.isfile(self._expt_config["save_file"]):
             self._log.warn("Clearing save file")
             os.remove(self._expt_config["save_file"])
 
@@ -108,11 +109,7 @@ class Experiment():
         if self._expt_config["n_processes"] == 1:
             return self._run_single()
         else:
-            result_list = self._run_multi()
-            result = pd.DataFrame()
-            for r in result_list:
-                result = pd.concat([result, r])
-            return result
+            return self._run_multi()
 
     def _run_single(self):
         """
@@ -123,10 +120,10 @@ class Experiment():
         self._log.warn("Run experiment in single thread")
         results = pd.DataFrame()
         for trial in self.__n_iterable(self._trial_config, self._expt_config["n_trials"]):
+            result = self._expt_config["experiment"](trial)
+            results = pd.concat([results, result])
             if self._expt_config["save_file"] is not None:
-                results = pd.concat([results, self._run_save(trial)])
-            else:
-                results = pd.concat([results, self._expt_config["experiment"](trial)])
+                self._save(result)
         return results
 
     def _run_multi(self):
@@ -136,24 +133,28 @@ class Experiment():
         :return: (pandas.DataFrame) data
         """
         self._log.warn("Run experiment in multiple processes")
+
+        results = pd.DataFrame()
         with Pool(self._expt_config["n_processes"]) as p:
-            if self._expt_config["save_file"] is not None:
-                return p.map(self._run_save, self._n_iterable(self._trial_config, self._expt_config["n_trials"]))
-            return p.map(self._expt_config["experiment"], self._n_iterable(self._trial_config, self._expt_config["n_trials"]))
+            run_map = p.imap(self._expt_config["experiment"], self.__n_iterable(self._trial_config, self._expt_config["n_trials"]))
+            for result in run_map:
+                if self._expt_config["save_file"] is not None:
+                    self._save(result) 
+                results = pd.concat([results, result])
 
-    def _run_save(self, trial_config):
+        return results
+
+    def _save(self, result):
         """
-        Run experiment and save results
+        Save results
 
-        :param trial_config: (list(dict)) Configurations for each trial
+        :param result: (pd.DataFrame) result to save
         :return: (pandas.DataFrame) data
         """
-        result = self._expt_config["experiment"](trial_config)
-        with self.__lock:
+        with expt_lock:
             data = import_file(self._expt_config["save_file"])
             data = pd.concat([data, result])
             export_file(data,self._expt_config["save_file"])
-        return result
 
     def __n_iterable(self, iterable_el, n = 1):
         """
